Refaktoring av notater
